Add service interfaces
- Move implementation to service implementations
- Adjust other classes to avoid circular dependencies
