+++ src/main/java/de/vinado/wicket/participate/providers/Select2PersonProvider.java
@@ -39,7 +39,7 @@ public class Select2PersonProvider extends ChoiceProvider<Person> {
     public Collection<Person> toChoices(final Collection<String> ids) {
         final ArrayList<Person> personList = new ArrayList<>();
         for (String id : ids) {
-            personList.add(personService.load(Person.class, Long.parseLong(id)));
+            personList.add(personService.getPerson(Long.parseLong(id)));
         }
         return personList;
     }
